# server/src/openapi/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Architecture Artifacts Core API",
    "description": "Core API endpoints for the Architecture Artifacts application - a comprehensive enterprise architecture documentation platform with multi-space file management, Git integration, authentication, templates, microservices, and search functionality.",
    "version": "2.0.0",
    "contact": {
      "name": "Design Artifacts Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Core API server"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session-based authentication using cookies"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "GitStatus": {
        "type": "object",
        "properties": {
          "ahead": {
            "type": "integer"
          },
          "behind": {
            "type": "integer"
          },
          "current": {
            "type": "string"
          },
          "tracking": {
            "type": "string"
          },
          "files": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "GitDraft": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "SyncStatus": {
        "type": "object",
        "properties": {
          "lastSync": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
...
    "/folders": {
      "post": {
        "summary": "Create new folder",
        "tags": ["Files"],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["folderPath"],
                "properties": {
                  "folderPath": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Folder created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/{space}/folders": {
      "post": {
        "summary": "Create folder in space",
        "tags": ["Space Files"],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "space",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Space name"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["folderPath"],
                "properties": {
                  "folderPath": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Folder created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Write access denied to space",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/commit": {
      "post": {
        "summary": "Commit changes to Git",
        "tags": ["Git"],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["message"],
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Changes committed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/clone": {
      "post": {
        "summary": "Clone Git repository",
        "tags": ["Git"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["repoUrl"],
                "properties": {
                  "repoUrl": {
                    "type": "string"
                  },
                  "branch": {
                    "type": "string",
                    "default": "main"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Repository cloned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "branch": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/pull": {
      "post": {
        "summary": "Pull changes from remote Git repository",
        "tags": ["Git"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "branch": {
                    "type": "string",
                    "default": "main"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Repository updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "branch": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/git-status": {
      "get": {
        "summary": "Get Git status for all spaces",
        "tags": ["Git"],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Git status for all accessible spaces",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/GitStatus"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/git/drafts": {
      "get": {
        "summary": "Get draft files",
        "tags": ["Git"],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of draft files",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GitDraft"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/git/discard-drafts": {
      "post": {
        "summary": "Discard draft changes",
        "tags": ["Git"],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Draft changes discarded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/git/git-status": {
      "get": {
        "summary": "Enhanced Git status with drafts",
        "tags": ["Git"],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Enhanced Git status information",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GitStatus"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "drafts": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/GitDraft"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/git/sync-status": {
      "get": {
        "summary": "Get last sync information",
        "tags": ["Git"],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sync status information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatus"
                }
              }
            }
          }
        }
      }
    },
    "/git/provider-info": {
      "get": {
        "summary": "Get filing provider information",
        "tags": ["Git"],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Filing provider information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "provider": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    },
                    "capabilities": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}